refactor: consolidate ChalanInfo component and update related imports
