+++ src/app/modules/transportManagement/routeCostManagement/rentalVehilceCost/helper.js
@@ -127,18 +127,6 @@ export const getSupplierDDL = async (accId, buId, setter) => {
   } catch (error) {}
 };
 
-// export const getChalanInfo = async (shipmentId, setter) => {
-//   try {
-//     const res = await Axios.get(
-//       `/tms/Shipment/GetChallanInfoByShipmentId?shipmentId=${shipmentId}`
-//     );
-
-//     if (res.status === 200) {
-//       setter(res?.data);
-//     }
-//   } catch (error) {}
-// };
-
 export const getShippingInfo = async (id, setter) => {
   try {
     const res = await Axios.get(
